Add support for macos

+++ install.sh
@@ -1,5 +1,20 @@
 #!/bin/bash
 
+DOTFILES=$(pwd)
+
+#######
+# zsh #
+#######
+
+mkdir -p "$XDG_CONFIG_HOME/zsh"
+ln -sf "$DOTFILES/zsh/.zshenv" "$HOME"
+ln -sf "$DOTFILES/zsh/.zshrc" "$XDG_CONFIG_HOME/zsh"
+
+ln -sf "$DOTFILES/zsh/aliases" "$XDG_CONFIG_HOME/zsh/aliases"
+
+rm -rf "$XDG_CONFIG_HOME/zsh/external"
+ln -sf "$DOTFILES/zsh/external" "$XDG_CONFIG_HOME/zsh"
+
 ########
 # nvim #
 ########
@@ -25,39 +40,35 @@ ln -sf "$DOTFILES/nvim/autoload/plug.vim" "$XDG_CONFIG_HOME/nvim/autoload/plug.v
 # Install (or update) all the plugins
 nvim --noplugin +PlugUpdate +qa
 
-######
-# i3 #
-######
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    ######
+    # i3 #
+    ######
 
-rm -rf "$XDG_CONFIG_HOME/i3"
-ln -s "$DOTFILES/i3" "$XDG_CONFIG_HOME"
+    rm -rf "$XDG_CONFIG_HOME/i3"
+    ln -s "$DOTFILES/i3" "$XDG_CONFIG_HOME"
 
-#######
-# zsh #
-#######
+    #########
+    # Fonts #
+    #########
 
-mkdir -p "$XDG_CONFIG_HOME/zsh"
-ln -sf "$DOTFILES/zsh/.zshenv" "$HOME"
-ln -sf "$DOTFILES/zsh/.zshrc" "$XDG_CONFIG_HOME/zsh"
+    mkdir -p "$XDG_DATA_HOME"
+    cp -rf "$DOTFILES/fonts" "$XDG_DATA_HOME"
 
-ln -sf "$DOTFILES/zsh/aliases" "$XDG_CONFIG_HOME/zsh/aliases"
+    #########
+    # Dunst #
+    #########
 
-rm -rf "$XDG_CONFIG_HOME/zsh/external"
-ln -sf "$DOTFILES/zsh/external" "$XDG_CONFIG_HOME/zsh"
+    mkdir -p "$XDG_CONFIG_HOME/dunst"
+    ln -sf "$DOTFILES/dunst/dunstrc" "$XDG_CONFIG_HOME/dunst/dunstrc"
+elif [[ "$OSTYPE" == "darwin"* ]]; then
+    ####################
+    # Alacritty Mac OS # 
+    ####################
 
-#########
-# Fonts #
-#########
-
-mkdir -p "$XDG_DATA_HOME"
-cp -rf "$DOTFILES/fonts" "$XDG_DATA_HOME"
-
-#########
-# Dunst #
-#########
-
-mkdir -p "$XDG_CONFIG_HOME/dunst"
-ln -sf "$DOTFILES/dunst/dunstrc" "$XDG_CONFIG_HOME/dunst/dunstrc"
+    mkdir -p "$XDG_CONFIG_HOME/alacritty"
+    ln -sf "$DOTFILES/alacritty/alacritty.macos.yml" "$XDG_CONFIG_HOME/alacritty/alacritty.yml"
+fi
 
 ########
 # tmux #
@@ -65,11 +76,4 @@ ln -sf "$DOTFILES/dunst/dunstrc" "$XDG_CONFIG_HOME/dunst/dunstrc"
 
 mkdir -p "$XDG_CONFIG_HOME/tmux"
 ln -sf "$DOTFILES/tmux/tmux.conf" "$XDG_CONFIG_HOME/tmux/tmux.conf"
-
-#############
-# Alacritty #
-#############
-
-mkdir -p "$XDG_CONFIG_HOME/alacritty"
-ln -sf "$DOTFILES/alacritty/alacritty.yml" "$XDG_CONFIG_HOME/alacritty/alacritty.yml"
